Split SendIRC, parse_irc_line, and get_ssl_content off into appropriate files
